backend/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lxml==4.5.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
pyrsistent==0.16.0
pytz==2020.1
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
